refactor(model): move inverse design out of the training class
FlexibleMultiTaskModel was 3296 lines, and 1536 of them were inverse design — target-driven
search over inputs, which is a *use* of a trained model rather than part of training one. It
arrived after the training code and grew in place. Nothing on the training path calls any of
it; the only consumers are workflows/inverse.py and workflows/inverse_trajectory.py.

The coupling turned out to be thin enough to make this nearly mechanical: those 1536 lines
reach for exactly three model attributes (encoder, task_heads, task_configs_map) and two
helpers (_head, _expand_for_kernel_regression), plus nn.Module's train/eval/training/parameters.
InverseDesignMixin declares that contract rather than assuming it — the first draft inherited
object and mypy immediately named all four module members it was silently borrowing, which is
the argument for declaring it.

A mixin, not free functions, because it keeps model.optimize_composition(...) working unchanged
across two workflow modules and 186 lines of tests, so this reads as one relocation rather than
a relocation plus an API change. Free functions taking the model explicitly would express the
boundary better and are a mechanical follow-up from here.

    flexible_multi_task_model.py   3083 -> 1475
    inverse_design.py                 0 -> 1699

Verified behaviour-preserving where it actually matters — every existing checkpoint has to keep
loading:

    state_dict keys   18 -> 18, identical
    parameter count   1265 -> 1265, identical
    MRO               FlexibleMultiTaskModel -> InverseDesignMixin -> LightningModule -> ...

536 tests pass. One decorator was nearly lost in the move: evaluate_targets carries
@torch.no_grad(), and slicing methods by their def line leaves decorators behind. That one is
semantic, not cosmetic — without it the targeted search builds a graph it never uses — so it
is restored explicitly rather than left to review to catch.
